+++ src/plot.h
@@ -21,7 +21,7 @@ void plotcorridor(struct corridor *c);
 void plottagwall(struct cube *c,int wallno,int hilight,int xor);
 void plotcurrent(void);
 void plotlevel(void);
-unsigned long cont_plotlevel(struct lightsource **ls);
+uint32_t cont_plotlevel(struct lightsource **ls);
 void copytoscreen(void);
 void initdescmap(void);
 void dec_frames(int ec);
